requirements.txt
django-notifications-hq==1.5.0
django-crispy-forms==1.9.1
django-messages==0.6.0
django-crispy-forms==1.9.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==5.0.1
django-cors-headers==3.5.0
dj-database-url==0.5.0
Django==3.0.6
django-storages==1.9.1
django-heroku==0.3.1
gunicorn==20.0.4
python-decouple==3.3
psycopg2==2.8.4
asgiref==3.2.3
django-utils-six==2.0
six==1.14.0

cloudinary==1.22.0
dj3-cloudinary-storage==0.0.3